Conceal presented message on logout
