+++ Eurofurence/Application/Components/No Content/NoContentPlaceholderViewController.swift
@@ -1,3 +1,10 @@
 import UIKit
 
-class NoContentPlaceholderViewController: UIViewController { }
+class NoContentPlaceholderViewController: UIViewController {
+    
+    static func fromStoryboard() -> NoContentPlaceholderViewController {
+        let storyboard = UIStoryboard(name: "NoContentPlaceholderViewController", bundle: nil)
+        return storyboard.instantiate(NoContentPlaceholderViewController.self)
+    }
+    
+}
