+++ Eurofurence/Application/Modules/Principal Content/PrincipalContentAggregator.swift
@@ -11,10 +11,9 @@ public struct PrincipalContentAggregator: PrincipalContentModuleProviding {
     public func makePrincipalContentModule() -> UIViewController {
         let applicationModules = applicationModuleFactories.map({ $0.makeApplicationModuleController() })
         let navigationControllers = applicationModules.map(NavigationController.init)
-        let noContentStoryboard = UIStoryboard(name: "NoContentPlaceholderViewController", bundle: nil)
         let splitViewControllers = navigationControllers.map { (navigationController) -> UISplitViewController in
             let splitViewController = SplitViewController()
-            let noContentPlaceholder = noContentStoryboard.instantiateInitialViewController().unsafelyUnwrapped
+            let noContentPlaceholder = NoContentPlaceholderViewController.fromStoryboard()
             splitViewController.viewControllers = [navigationController, noContentPlaceholder]
             splitViewController.tabBarItem = navigationController.tabBarItem
